INT-2772 use integer division

+++ fn_xforce/fn_xforce/components/xforce_get_collection_by_id.py
@@ -79,7 +79,7 @@ class FunctionComponent(ResilientComponent):
                     "get", request_string, proxies=proxies, auth=(XFORCE_APIKEY, XFORCE_PASSWORD), callback=helper.handle_case_response)
                 # Is the status code in the 2XX family?
                 # Save returned case files
-                if res.status_code / 100 == 2:
+                if int(res.status_code / 100) == 2:
                     case_files = res.json()
             except Exception:
                 raise ValueError("Encountered issue when contacting XForce API")
